requirements.txt
streamlit
pandas
google-generativeai
selenium
numpy
torch==2.0.1+cpu
matplotlib
requests
bs4
sentence-transformers==1.2.1
transformers==4.36.2
python-dotenv
fastapi
uvicorn
